Simplify DI types around Resource
