--- src/DI/Dependency.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Utopia\DI;
-
-class Dependency extends Injection
-{
-}
